refactor: delete data.js
