+++ js/create-offers.js
@@ -1,5 +1,39 @@
 import {getInteger, getFloatingPointNumber, getRandomArrayElement, createRandomArr} from './util.js';
-import {TYPE, CHECKIN, CHECKOUT, FEATURES, PHOTOS} from './data.js';
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const  FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
 
 const OFFERS_AMOUNT = 10;
 
@@ -38,3 +72,5 @@ const createAd = (index) => {
 const getAds = (adsCount) => Array.from({length: adsCount}, (item, index) => createAd(index + 1));
 
 getAds(OFFERS_AMOUNT);
+
+export {getAds, OFFERS_AMOUNT};
