+++ js/main.js
@@ -1,1 +1,4 @@
-import './create-offers.js';
+/* eslint-disable no-undef */
+import {getAds, OFFERS_AMOUNT} from './create-offers.js';
+
+getAds(OFFERS_AMOUNT);
